transports: add callCustomManifestLoader function
